requirements.txt
BeautifulSoup==3.2.1
certifi==2017.4.17
chardet==3.0.3
idna==2.5
python-dateutil==2.6.0
pytz==2017.2
requests==2.17.3
six==1.10.0
urllib3==1.21.1
